refactor: DEFI-2810: Colocate stablecoin-symbol recorder with its caller
PR 1 (50ae2ae) put `record_per_forex_metrics` next to its only caller
`update_forex_store`, in `periodic.rs`. PR 2's `record_exchange_outcome`
follows the same rule — it lives in `lib.rs` next to `call_exchange`,
which is where it's invoked from. `record_stablecoin_symbol_rates_received`
was the outlier: defined in `lib.rs` but called only from `api.rs`'s
`get_stablecoin_rate`. Moves the function (and its three unit tests) to
`api.rs` so all three per-source recorders follow the same locality
rule and the next reader doesn't have to wonder why one of them sits
two files away from its single caller.

No visibility change beyond what the move requires: the function drops
to module-private (the only caller is now in the same file), and
`api.rs` gains direct imports of `add_labeled_counter`, `LabelKey`, and
`MetricName` to match.
